+++ cdef.php
@@ -168,19 +168,16 @@ function form_actions() {
 	
 	/* if we are to save this form, instead of display it */
 	if (isset($_POST['selected_items'])) {
-		$selected_items = unserialize(stripslashes($_POST['selected_items']));
+		$selected_items = sanitize_unserialize_selected_items($_POST['selected_items']);
 
-		if ($_POST['drp_action'] == '1') { /* delete */
-			db_execute('DELETE FROM cdef WHERE ' . array_to_sql_or($selected_items, 'id'));
-			db_execute('DELETE FROM cdef_items WHERE ' . array_to_sql_or($selected_items, 'cdef_id'));
-
-		}elseif ($_POST['drp_action'] == '2') { /* duplicate */
-			for ($i=0;($i<count($selected_items));$i++) {
-				/* ================= input validation ================= */
-				input_validate_input_number($selected_items[$i]);
-				/* ==================================================== */
-
-				duplicate_cdef($selected_items[$i], $_POST['title_format']);
+		if ($selected_items != false) {
+			if ($_POST['drp_action'] == '1') { /* delete */
+				db_execute('DELETE FROM cdef WHERE ' . array_to_sql_or($selected_items, 'id'));
+				db_execute('DELETE FROM cdef_items WHERE ' . array_to_sql_or($selected_items, 'cdef_id'));
+			}elseif ($_POST['drp_action'] == '2') { /* duplicate */
+				for ($i=0;($i<count($selected_items));$i++) {
+					duplicate_cdef($selected_items[$i], $_POST['title_format']);
+				}
 			}
 		}
